Put Setting Groups into extra files
